Add option to write the delimiter explicitly in CSV
Hi, Excel shows a different version of the .csv file in different regions.
On some computers it's better to use the comma, on others the semicolon. 

By explicitly writing the delimiter we avoid this problem. 

I know it seems like a Excel specific fix, but Excel is very popular and I made intentionaly the parameter optional so we avoid the necessity to specify it.

https://stackoverflow.com/questions/17953679/how-to-correctly-display-csv-files-within-excel-2013#19769403